third_party/WebKit/LayoutTests/media/media-controller-playback-expected.txt

 
RUN(controller = video.controller)
EVENT(canplaythrough)
TEST(controller.currentTime = -Infinity) THROWS("TypeError: Failed to set the 'currentTime' property on 'MediaController': The provided double value is non-finite.") OK
TEST(controller.currentTime = Infinity) THROWS("TypeError: Failed to set the 'currentTime' property on 'MediaController': The provided double value is non-finite.") OK
TEST(controller.currentTime = NaN) THROWS("TypeError: Failed to set the 'currentTime' property on 'MediaController': The provided double value is non-finite.") OK
RUN(controller.play())
EVENT(playing)
EXPECTED (controller.paused == 'false') OK
RUN(controller.currentTime = 5)
EXPECTED (controller.currentTime == '5') OK
EXPECTED (video.currentTime == '5') OK
EXPECTED (video2.currentTime == '5') OK
EVENT(ended)
END OF TEST
